refactor: update import paths to use crate::dom for stream-related modules
