+++ components/script/dom/stream/readablestreamgenericreader.rs
@@ -8,7 +8,7 @@ use js::jsval::UndefinedValue;
 use js::rust::HandleValue as SafeHandleValue;
 
 use super::readablestream::ReaderType;
-use super::types::ReadableStream;
+use crate::dom::types::ReadableStream;
 use crate::dom::bindings::error::{Error, ErrorToJsval, Fallible};
 use crate::dom::bindings::reflector::DomGlobal;
 use crate::dom::bindings::root::{DomRoot, MutNullableDom};
