+++ components/script/dom/stream/transformstream.rs
@@ -17,13 +17,13 @@ use rustc_hash::FxHashMap;
 use script_bindings::callback::ExceptionHandling;
 use script_bindings::realms::InRealm;
 
-use super::bindings::codegen::Bindings::QueuingStrategyBinding::QueuingStrategySize;
-use super::bindings::structuredclone::StructuredData;
-use super::bindings::transferable::Transferable;
-use super::messageport::MessagePort;
-use super::promisenativehandler::Callback;
+use crate::dom::bindings::codegen::Bindings::QueuingStrategyBinding::QueuingStrategySize;
+use crate::dom::bindings::structuredclone::StructuredData;
+use crate::dom::bindings::transferable::Transferable;
+use crate::dom::messageport::MessagePort;
+use crate::dom::promisenativehandler::Callback;
 use super::readablestream::CrossRealmTransformReadable;
-use super::types::{TransformStreamDefaultController, WritableStream};
+use crate::dom::types::{TransformStreamDefaultController, WritableStream};
 use super::writablestream::CrossRealmTransformWritable;
 use crate::dom::bindings::cell::DomRefCell;
 use crate::dom::bindings::codegen::Bindings::QueuingStrategyBinding::QueuingStrategy;
